added 'Fixed within the Browser' section
Far away from finished, this was only what comes to my mind. I will somehow integrate this directly to the specific sections so that if you e.g. click on the index (todo) you get e.g. HTML5 and then see what exactly is maybe vulnerable or known as 'insecure'.