Several papers added First good paper I found for extension 'security' and a more complex analasis. In fact it's the evidence that faking or spoofind data in extensions like eg. Timing side channel attacks.
From the paper:
"In particular, they changed the initial scheme (moz-extension://[extID]/[path]) to moz-extension://[random-UUID]/[path]. Unfortunately, while this change makes indeed more difficult to enumerate user extensions, it introduces a far more dangerous problem. In fact, the random UUID token can now be used to precisely fingerprint users if it is leaked by an extensions. A website can retrieve this UUID and use it to uniquely identify the user, as once it is generated the random ID never changes.

Also as conclusion:
"In this paper, we show that the current countermeasures adopted by all browser families are insufficient or erroneously implemented."

Firefox in fact has more problem, their anti-fingerprinting mechanism are in 80% vulnerable, wrongly implemented or can be bypassed as shown in this paper. The problem I see here is that since all Browsers are depending on WebExtensions makes it a larger attack surface, since an attacker 'only' needs to abuse the known weaknesses, which then affects a larger scale of Browsers. Chrome(ium) was in this paper more vulnerable to side-channel attacks while Firefox and Edge are more attackable from wrongly implemented 'fuzzing' tricks.

The entire extension problematic is imho underrated cause there in all stores I've seen tons of fake applications just search for 'ublock' and you'll find 10 fakes at minimum in every store under changed name with changed icons but overall the same ripp-off repacked with malware.

Signed-off-by: CHEF-KOCH <CHEF-KOCH@protonmail.com>